Move logging and time formatting to log.py, and fix imports
Works towards #130.
