+++ privcount/data_collector.py
@@ -9,12 +9,13 @@ from time import time
 from copy import deepcopy
 from base64 import b64decode
 
-from protocol import PrivCountClientProtocol, TorControlClientProtocol
-from counter import SecureCounters, counter_modulus, add_counter_limits_to_config, combine_counters
-from util import log_error, get_public_digest_string, load_public_key_string, encrypt, format_delay_time_wait, format_last_event_time_since, normalise_path, choose_secret_handshake_path, PrivCountClient
-
 from twisted.internet import task, reactor, ssl
 from twisted.internet.protocol import ReconnectingClientFactory
+
+from privcount.counter import SecureCounters, counter_modulus, add_counter_limits_to_config, combine_counters
+from privcount.log import log_error, format_delay_time_wait, format_last_event_time_since
+from privcount.protocol import PrivCountClientProtocol, TorControlClientProtocol
+from privcount.util import get_public_digest_string, load_public_key_string, encrypt, normalise_path, choose_secret_handshake_path, PrivCountClient
 
 # using reactor: pylint: disable=E1101
 # method docstring missing: pylint: disable=C0111
